+++ lib/couch_potato/rspec/matchers.rb
@@ -1,0 +1,2 @@
+require File.dirname(__FILE__) + '/matchers/map_to_matcher'
+require File.dirname(__FILE__) + '/matchers/reduce_to_matcher'