+++ lib/couch_potato/rspec/matchers/reduce_to_matcher.rb
@@ -1,0 +1,6 @@
+module CouchPotato
+  module RSpec
+    class ReduceToMatcher
+    end
+  end
+end